# src/lib/openapi.yaml
openapi: 3.0.0
info:
  title: PetStayTion API
  version: 1.0.0
  description: API for pet care service management
servers:
  - url: /
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/bookings/service-provider/notifications:
    get:
      summary: Get recent booking notifications for service provider
      description: |
        Retrieves pending bookings (always) and recent completed/cancelled/rejected bookings within time windows for the authenticated service provider.
        - PENDING: No time limit (always returned).
        - CANCELLED/REJECTED: Only if updated within the last 1 month.
        - ACCEPTED: No time limit (to check for status updates).
        - AWAITING_PAYMENT: No time limit.
        - OVERDUE: Only if updated within the last 3 months.
        - PAID: Only if updated within the last 3 months.
        Automatically updates booking status to AWAITING_PAYMENT if end date has passed, and to OVERDUE if 48h after end date without payment.
        Requires a valid Firebase authentication token.
        Only available to active service providers.
        Returns details about all pets in the booking, client (name, email, phone), message, booking times, status, and advertisement ID.
      tags:
        - Bookings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Recent bookings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  bookings:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        status:
                          type: string
                          enum:
                            - PENDING
                            - ACCEPTED
                            - REJECTED
                            - CANCELLED
                            - AWAITING_PAYMENT
                            - OVERDUE
                            - PAID
                          example: "PENDING"
                        startDateTime:
                          type: string
                          format: date-time
                          example: "2025-10-25T10:00:00Z"
                        endDateTime:
                          type: string
                          format: date-time
                          example: "2025-10-25T12:00:00Z"
                        message:
                          type: string
                          nullable: true
                          example: "Please confirm the exact hours."
                        advertisementId:
                          type: integer
                          nullable: true
                          example: 123
                        pets:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                                example: 1
                              name:
                                type: string
                                example: "Max"
                              age:
                                type: number
                                example: 5
                              description:
                                type: string
                                nullable: true
                                example: "Friendly dog"
                              chronicDiseases:
                                type: array
                                items:
                                  type: string
                                example:
                                  - "Allergies"
                              isHealthy:
                                type: boolean
                                nullable: true
                                example: true
                              species:
                                type: string
                                example: "Dog"
                              keyImage:
                                type: string
                                nullable: true
                                example: "https://example.com/pet.jpg"
                        client:
                          type: object
                          properties:
                            firstName:
                              type: string
                              nullable: true
                              example: "John"
                            lastName:
                              type: string
                              nullable: true
                              example: "Doe"
                            email:
                              type: string
                              nullable: true
                              example: "john.doe@example.com"
                            phoneNumber:
                              type: string
                              nullable: true
                              example: "123456789"
        "401":
          description: Unauthorized (invalid or missing token)
        "403":
          description: Forbidden (user is not an active service provider)
        "404":
          description: User not found
        "500":
          description: Internal server error
  /api/bookings/service-provider/reject:
    post:
      summary: Reject a pending booking as service provider
      description: |
        Allows the authenticated service provider to reject a pending booking.
        Only pending bookings can be rejected.
        Requires a valid Firebase authentication token.
        Only available to active service providers.
      tags:
        - Bookings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingId:
                  type: integer
                  example: 1
                  description: The ID of the booking to reject
      responses:
        "200":
          description: Booking rejected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Booking rejected successfully"
        "400":
          description: Bad request (invalid booking ID or not pending)
        "401":
          description: Unauthorized (invalid or missing token)
        "403":
          description: Forbidden (user is not an active service provider or unauthorized for this booking)
        "404":
          description: Booking or user not found
        "500":
          description: Internal server error
...
  /api/pets/{id}/diagnoses:
    get:
      summary: Get all diagnostic analyses for a specific pet
      description: Returns a list of all AI diagnostic analyses performed for the given pet. Only accessible by the pet owner.
      tags: [Analysis]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Pet ID
      responses:
        "200":
          description: List of diagnostic analyses
          content:
            application/json:
              schema:
                type: object
                properties:
                  analyses:
                    type: array
                    items:
                      type: object
                      properties:
                        idAnalysis:
                          type: integer
                          example: 42
                        createdAt:
                          type: string
                          format: date-time
                          example: "2025-11-18T10:30:00.000Z"
                        overallHealth:
                          type: string
                          enum: [healthy, hard to tell, unhealthy]
                          example: "healthy"
        "400":
          description: Invalid pet ID
        "401":
          description: Unauthorized (missing or invalid token)
        "403":
          description: Forbidden (not the pet owner)
        "404":
          description: Pet not found
        "500":
          description: Internal server error
...